Add 'assert only' arg to check files & not rewrite
Add a arg field and config value to allow checks on your code base but
not re-write files if they are not compliant.

`importanize` can then be added to CI checks to assert that the code
base is compliant and fail if not correct. `importanize` will exit
with status code `1` if this flag is enabled for a run and the code base
checked needs fixing.